pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "task_manager"
requires-python = ">=3.10"
dependencies = [
    "django>=5.0",
    "gunicorn>=21.2.0",
    "python-dotenv>=1.0.0",
    "dj-database-url>=2.1.0",
    "psycopg2-binary>=2.9.9",
    "django-bootstrap5>=26.2",
    "django-filter>=25.2",
]
